# Simple: snarkworker log and times