--- a/src/lib/snark_worker_lib/worker.ml
+++ b/src/lib/snark_worker_lib/worker.ml
@@ -107,18 +107,18 @@ module Make (Inputs : Intf.Inputs_intf) :
             Worker_state.worker_wait_time
             +. (0.5 *. Random.float Worker_state.worker_wait_time)
           in
-          Logger.trace log "No work received. Sleeping %.3f seconds."
-            random_delay ;
+          Logger.trace log "No work received. Sleeping %.4fs" random_delay ;
           let%bind () = wait ~sec:random_delay () in
           go ()
       | Ok (Some work) -> (
           Logger.info log !"Received work.%!" ;
+          Core.(flush stdout) ;
           match perform state public_key work with
           | Error e -> log_and_retry "performing work" e
           | Ok result -> (
               match%bind
-                Logger.info log !"Submitting work.%!" ;
                 emit_proof_metrics result.metrics log ;
+                Logger.info log !"Submitting work.%!" ;
                 dispatch Rpcs.Submit_work.rpc shutdown_on_disconnect result
                   daemon_address
               with