--- a/src/lib/snark_worker_lib/worker.ml
+++ b/src/lib/snark_worker_lib/worker.ml
@@ -112,7 +112,7 @@ module Make (Inputs : Intf.Inputs_intf) :
           go ()
       | Ok (Some work) -> (
           Logger.info log !"Received work.%!" ;
-          let%bind `This_is_async__Think_about_blocking =
+          let `This_is_async__Think_about_blocking =
             Async.flush `This_is_async__Think_about_blocking
           in
           (* Pause to flush stdout *)