Use the pre-existing metrics information to emit some metrics at runtime about snark performance.

- Added explicit logs at Receipt and Submission of work.
- Added information exposing type of work found and time spent processing.
- Change default sleep from 10s to 5s with more splay (random deviation from 20% to 50%)
- Increased sleep when unable to connect from 0.5s to 30s 

Before:
```
{...,"message":"No work; waiting 11.447s"}
* Elements of w skipped: 8350 (7.88%)
* Elements of w processed with special addition: 5888 (5.55%)
* Elements of w remaining: 91766 (86.57%)
* Elements of w skipped: 8078 (16.99%)
* Elements of w processed with special addition: 5747 (12.09%)
<...>
* Proof size in bits: 1195
* G1 elements in proof: 2
* G2 elements in proof: 1
* Proof size in bits: 1493
{...,"message":"Received work."}
{...,"message":"No work; waiting 10.853s"}
```

After:
```
{...,"message":"No work received from 127.0.0.1:8501 - sleeping 7.3979s"}
{...,"message":"Received work from 127.0.0.1:8501"}
* Elements of w skipped: 70682 (30.15%)
* Elements of w processed with special addition: 55604 (23.72%)
* Elements of w remaining: 108113 (46.12%)
<...>
* G1 elements in proof: 2
* G2 elements in proof: 1
* Proof size in bits: 1493
{...,"message":"Base Proof Completed - 41.4024s"}
{...,"message":"Submitted work to 127.0.0.1:8501"}
```