@jkrauska 2019-03-08
If wondering why I added host/port.  In the future we can generalize the snark-worker to service MANY different test nets concurrently by polling on a set of host/ips. (as long as the snark circuit is fixed)